Renaming helloworld feature to home Removing sample string and replacing with app name Removing visibility disposable (to be added back in example repo)
