+++ app/src/main/java/com/ustwo/boilerplate/feature/home/HomeComponent.java
@@ -1,4 +1,4 @@
-package com.ustwo.boilerplate.feature.helloworld;
+package com.ustwo.boilerplate.feature.home;
 
 import com.ustwo.boilerplate.application.ApplicationComponent;
 import com.ustwo.boilerplate.base.PerViewScope;
@@ -6,6 +6,6 @@ import dagger.Component;
 
 @PerViewScope
 @Component(dependencies = ApplicationComponent.class)
-interface HelloWorldComponent {
-    void inject(HelloWorldActivity activity);
+interface HomeComponent {
+    void inject(HomeActivity activity);
 }
