+++ app/src/main/java/com/ustwo/boilerplate/feature/home/HomePresenter.java
@@ -1,4 +1,4 @@
-package com.ustwo.boilerplate.feature.helloworld;
+package com.ustwo.boilerplate.feature.home;
 
 import com.ustwo.boilerplate.base.BasePresenter;
 import com.ustwo.boilerplate.base.BaseView;
@@ -6,10 +6,10 @@ import com.ustwo.boilerplate.base.PerViewScope;
 import javax.inject.Inject;
 
 @PerViewScope
-class HelloWorldPresenter extends BasePresenter<HelloWorldPresenter.View> {
+class HomePresenter extends BasePresenter<HomePresenter.View> {
 
     @Inject
-    HelloWorldPresenter() {
+    HomePresenter() {
     }
 
     interface View extends BaseView {
