+++ app/src/main/res/layout/activity_home.xml
@@ -4,11 +4,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.ustwo.boilerplate.feature.helloworld.HelloWorldActivity">
+    tools:context="com.ustwo.boilerplate.feature.home.HomeActivity">
 
     <TextView
-        android:id="@+id/textView_helloWorld"
+        android:id="@+id/textView_appName"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="@string/hello_world"/>
+        android:text="@string/app_name"/>
 </RelativeLayout>
