+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,3 @@
 <resources>
     <string name="app_name">Boilerplate</string>
-    <string name="hello_world">Hello world</string>
 </resources>
